requirements.txt
aiogram>=3.27.0
python-dotenv>=1.0.1
playwright>=1.54.0
google-genai>=1.10.0
pydantic>=2.10.0
matplotlib>=3.9.0
